requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.10.6
pydantic-settings==2.7.1
email-validator>=2.0.0

# Async & WebSocket
websockets==12.0
python-multipart==0.0.6
httpx==0.26.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# Database & ORM
sqlalchemy>=2.0.36  # Python 3.13 compatible
alembic>=1.13.1
psycopg2-binary==2.9.10  # PostgreSQL driver (Python 3.13 compatible)

# Vector Database & RAG (pgvector via psycopg2)
sentence-transformers==2.3.1
langchain==0.1.4
langchain-community==0.0.16

# LLM Integration
openai>=1.10.0
tiktoken>=0.8.0
cohere>=4.45  # Alternative/fallback LLM

# Caching
redis==5.0.1
hiredis==2.3.2  # Faster Redis parser

# Rate Limiting
slowapi==0.1.9

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.3

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.40.0  # Error tracking (optional)

# Utilities
python-slugify==8.0.2
pendulum==3.0.0  # Better datetime handling
tenacity==8.2.3  # Retry logic
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For async testing
faker==22.5.1

# Development Tools
black==24.1.1
ruff==0.1.14
mypy==1.8.0
pre-commit==3.6.0

# Production Server
gunicorn==21.2.0
pgvector==0.4.1
